refactor: remove broken Nivo-based RisingStarsChart and use custom implementation
- Deleted the non-working Nivo-based RisingStarsChart component
- Renamed RisingStarsChartSimple to RisingStarsChart
- Kept the working custom implementation with avatar bubbles
- Updated all references and stories to use the renamed component

This simplifies the codebase by removing the problematic Nivo dependency
that was causing "Cannot read properties of undefined" errors.
