+++ src/components/features/analytics/RisingStarsChart.stories.tsx
@@ -19,33 +19,20 @@ const generateMockData = (count: number = 30): RisingStarsData[] => {
   const contributors = Array.from({ length: count }, (_, i) => {
     const isNewContributor = Math.random() > 0.7;
     const isRisingStar = Math.random() > 0.8;
-
-    const commits = Math.floor(Math.random() * 50) + 5;
-    const pullRequests = Math.floor(Math.random() * 20) + 2;
-    const issues = Math.floor(Math.random() * 10);
-    const comments = Math.floor(Math.random() * 40) + 5;
-    const reviews = Math.floor(Math.random() * 15);
-    const discussions = Math.floor(Math.random() * 8);
-    const totalGithubEvents = commits + pullRequests + issues + comments + reviews + discussions;
-    const codeContributions = commits + pullRequests;
-    const nonCodeContributions = issues + comments + reviews + discussions;
+    const baseActivity = Math.floor(Math.random() * 20) + 1;
 
     return {
-      x: codeContributions, // Code contributions (PRs + commits)
-      y: nonCodeContributions, // Non-code contributions
+      x: Math.floor(Math.random() * 50) + 5, // commits
+      y: Math.floor(Math.random() * 30) + 2, // PRs + issues
       size: Math.random() * 80 + 20, // velocity score scaled
       contributor: {
         login: `contributor-${i + 1}`,
         avatar_url: `https://avatars.githubusercontent.com/u/${Math.floor(Math.random() * 100000)}`,
         github_id: Math.floor(Math.random() * 100000),
-        commits,
-        pullRequests,
-        issues,
-        comments,
-        reviews,
-        discussions,
-        totalGithubEvents,
-        totalActivity: totalGithubEvents, // for compatibility
+        commits: Math.floor(Math.random() * 50) + 5,
+        pullRequests: Math.floor(Math.random() * 20) + 2,
+        issues: Math.floor(Math.random() * 10),
+        totalActivity: baseActivity,
         velocityScore: Math.random() * 10 + 1,
         growthRate: isRisingStar ? Math.random() * 200 + 50 : Math.random() * 50,
         firstContributionDate: new Date(
@@ -79,7 +66,7 @@ export const Default: Story = {
 
 export const ManyContributors: Story = {
   args: {
-    data: generateMockData(100),
+    data: generateMockData(50),
     height: 600,
     maxBubbles: 50,
   },
@@ -90,117 +77,6 @@ export const FewContributors: Story = {
     data: generateMockData(10),
     height: 400,
     maxBubbles: 50,
-  },
-};
-
-export const HighActivity: Story = {
-  args: {
-    data: [
-      {
-        id: 'rising-stars',
-        data: Array.from({ length: 20 }, (_, i) => {
-          const commits = Math.floor(Math.random() * 30) + 50;
-          const pullRequests = Math.floor(Math.random() * 15) + 20;
-          const issues = Math.floor(Math.random() * 10) + 10;
-          const comments = Math.floor(Math.random() * 60) + 40;
-          const reviews = Math.floor(Math.random() * 25) + 10;
-          const discussions = Math.floor(Math.random() * 15) + 5;
-          const totalGithubEvents =
-            commits + pullRequests + issues + comments + reviews + discussions;
-          const codeContributions = commits + pullRequests;
-          const nonCodeContributions = issues + comments + reviews + discussions;
-
-          return {
-            x: codeContributions, // Code contributions
-            y: nonCodeContributions, // Non-code contributions
-            size: Math.random() * 50 + 50, // larger bubbles
-            contributor: {
-              login: `power-user-${i + 1}`,
-              avatar_url: `https://avatars.githubusercontent.com/u/${i + 1000}`,
-              github_id: i + 1000,
-              commits,
-              pullRequests,
-              issues,
-              comments,
-              reviews,
-              discussions,
-              totalGithubEvents,
-              totalActivity: totalGithubEvents,
-              velocityScore: Math.random() * 5 + 15,
-              growthRate: Math.random() * 300 + 100,
-              firstContributionDate: new Date(
-                Date.now() - Math.random() * 180 * 24 * 60 * 60 * 1000
-              ).toISOString(),
-              lastContributionDate: new Date().toISOString(),
-              contributionSpan: Math.floor(Math.random() * 180) + 1,
-              isNewContributor: false,
-              isRisingStar: true,
-            },
-          };
-        }),
-      },
-    ],
-    height: 500,
-  },
-};
-
-export const NewContributorsWave: Story = {
-  args: {
-    data: [
-      {
-        id: 'rising-stars',
-        data: Array.from({ length: 25 }, (_, i) => {
-          const isNew = i < 15;
-          const commits = Math.floor(Math.random() * 20) + (isNew ? 5 : 15);
-          const pullRequests = Math.floor(Math.random() * 10) + (isNew ? 2 : 8);
-          const issues = Math.floor(Math.random() * 5) + (isNew ? 1 : 3);
-          const comments = Math.floor(Math.random() * 30) + (isNew ? 5 : 20);
-          const reviews = Math.floor(Math.random() * 10) + (isNew ? 2 : 8);
-          const discussions = Math.floor(Math.random() * 5) + (isNew ? 1 : 3);
-          const totalGithubEvents =
-            commits + pullRequests + issues + comments + reviews + discussions;
-          const codeContributions = commits + pullRequests;
-          const nonCodeContributions = issues + comments + reviews + discussions;
-
-          return {
-            x: codeContributions,
-            y: nonCodeContributions,
-            size: Math.random() * 60 + 30,
-            contributor: {
-              login: `${isNew ? 'new' : 'active'}-contributor-${i + 1}`,
-              avatar_url: `https://avatars.githubusercontent.com/u/${i + 5000}`,
-              github_id: i + 5000,
-              commits,
-              pullRequests,
-              issues,
-              comments,
-              reviews,
-              discussions,
-              totalGithubEvents,
-              totalActivity: totalGithubEvents,
-              velocityScore: Math.random() * 8 + (isNew ? 2 : 5),
-              growthRate: isNew ? Math.random() * 500 + 200 : Math.random() * 100,
-              firstContributionDate: new Date(
-                Date.now() - (isNew ? 30 : 200) * 24 * 60 * 60 * 1000
-              ).toISOString(),
-              lastContributionDate: new Date().toISOString(),
-              contributionSpan: isNew ? 30 : 200,
-              isNewContributor: isNew,
-              isRisingStar: isNew && i < 5,
-            },
-          };
-        }),
-      },
-    ],
-    height: 500,
-  },
-};
-
-export const CompactView: Story = {
-  args: {
-    data: generateMockData(15),
-    height: 350,
-    maxBubbles: 15,
   },
 };
